Fix compilation errors and warnings
Fix after 75a7a93db4d153528cca211a341d83c0bb8fdf24.

+++ apps/PQCgenKAT_sign.c
@@ -265,7 +265,7 @@ int ReadHex(FILE *infile, unsigned char *A, int Length, char *str) {
   return 1;
 }
 
-void fprintBstr(FILE *fp, char *S, unsigned char *A, unsigned long long L) {
+void fprintBstr(FILE *fp, char *S, unsigned char *A, size_t L) {
   unsigned long long i;
 
   fprintf(fp, "%s", S);
